backend/app/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
django-cors-headers==4.7.0